python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-vfs"
version = "0.1.0"
description = "Persistent virtual filesystem for AI agents, backed by SQLite or Postgres"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["filesystem", "ai", "agents", "sqlite", "postgres", "memory", "tools", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9"]
test = ["pytest>=8.0"]

[project.urls]
Homepage = "https://github.com/johannesmichalke/agent-vfs"
Repository = "https://github.com/johannesmichalke/agent-vfs"

[tool.pytest.ini_options]
testpaths = ["tests"]
